add noop flag

+++ ceph_rsnapshot/dirs.py
@@ -38,23 +38,32 @@ def setup_temp_conf_dir(pool=''):
       logger.info('using temp conf dir %s' % settings.TEMP_CONF_DIR)
     else:
       try:
-        os.makedirs(settings.TEMP_CONF_DIR)
-        logger.info('created temp dir at %s' % settings.TEMP_CONF_DIR)
+        if settings.NOOP:
+          logger.info('NOOP: would have made temp dir %s' % settings.TEMP_CONF_DIR)
+        else:
+          os.makedirs(settings.TEMP_CONF_DIR)
+          logger.info('created temp dir at %s' % settings.TEMP_CONF_DIR)
       except Exception as e:
         logger.error('Cannot create conf temp dir (or intermediate dirs) from'+
         ' setting %s with error %s' % (settings.TEMP_CONF_DIR, e))
         sys.exit(1)
   else:
     try:
-      temp_conf_dir = tempfile.mkdtemp(prefix=settings.TEMP_CONF_DIR_PREFIX)
-      logger.info('creating temp conf dir: %s' % temp_conf_dir)
+      if settings.NOOP:
+        logger.info('NOOP: would have made temp dir with mkdtemp prefix' % settings.TEMP_CONF_DIR_PREFIX)
+      else:
+        temp_conf_dir = tempfile.mkdtemp(prefix=settings.TEMP_CONF_DIR_PREFIX)
+        logger.info('created temp conf dir: %s' % temp_conf_dir)
       # store this in global settings
       settings.TEMP_CONF_DIR = temp_conf_dir
     except Exception as e:
       logger.error('cannot create conf temp dir with error %s' % e)
       sys.exit(1)
-  logger.info('creating temp conf subdir for pool %s' % pool)
-  os.mkdir("%s/%s" % (settings.TEMP_CONF_DIR,pool), 0700)
+  if settings.NOOP:
+    logger.info('NOOP: would have made temp conf subdir for pool %s' % pool)
+  else:
+    logger.info('creating temp conf subdir for pool %s' % pool)
+    os.mkdir("%s/%s" % (settings.TEMP_CONF_DIR,pool), 0700)
   return settings.TEMP_CONF_DIR
 
 # make path to export qcows to
@@ -65,8 +74,11 @@ def setup_qcow_temp_path(pool=''):
   logger = logs.get_logger()
   temp_path = settings.QCOW_TEMP_PATH
   if not os.path.isdir("%s/%s" % (temp_path,pool)):
-    logger.info('creating qcow temp path: %s/%s' % (temp_path,pool))
-    os.makedirs("%s/%s" % (temp_path,pool),0700)
+    if settings.NOOP:
+      logger.info('NOOP: would have made qcow temp path %s/%s' % (temp_path,pool))
+    else:
+      logger.info('creating qcow temp path: %s/%s' % (temp_path,pool))
+      os.makedirs("%s/%s" % (temp_path,pool),0700)
   else:
     logger.info('using qcow temp path: %s/%s' % (temp_path,pool))
 
@@ -82,7 +94,10 @@ def make_empty_source():
     if len(dirlist) != 0:
       raise NameError('temp_path_not_empty')
   except:
-    os.mkdir(temp_path,0700)
+    if settings.NOOP:
+      logger.info('NOOP: would have made temp empty source at %s' % temp_path)
+    else:
+      os.mkdir(temp_path,0700)
     # TODO catch if error
 
 def setup_dir(directory):
@@ -90,7 +105,10 @@ def setup_dir(directory):
   # make the if it doesn't exist
   if not os.path.isdir(directory):
     # make dir and preceeding dirs if necessary
-    os.makedirs(directory,0700)
+    if settings.NOOP:
+      logger.info('NOOP: would have made directory %s' % directory)
+    else:
+      os.makedirs(directory,0700)
 
 def setup_dir_per_pool(directory):
   logger = logs.get_logger()
@@ -104,8 +122,12 @@ def remove_temp_conf_dir():
   if not settings.KEEPCONF:
     logger.info("removing temp conf dir %s" % settings.TEMP_CONF_DIR)
     try:
-      # TODO all pools
-      os.rmdir("%s/%s" % (settings.TEMP_CONF_DIR, settings.POOL))
-      os.rmdir(settings.TEMP_CONF_DIR)
+      if settings.NOOP:
+        logger.info('would have removed temp conf dirs %s/%s and %s',
+          (settings.TEMP_CONF_DIR, settings.POOL, settings.TEMP_CONF_DIR))
+      else:
+        # TODO all pools
+        os.rmdir("%s/%s" % (settings.TEMP_CONF_DIR, settings.POOL))
+        os.rmdir(settings.TEMP_CONF_DIR)
     except Exception as e:
       logger.warning("unable to remove temp conf dir with error %s" % e)
