+++ ceph_rsnapshot/utils/remove_qcow.py
@@ -13,6 +13,7 @@ def remove_qcow():
                                    argument_default=argparse.SUPPRESS)
   parser.add_argument('image')
   parser.add_argument('--pool', required=False)
+  parser.add_argument("--noop", action='store_true',required=False, help="noop - don't make any directories or do any actions. logging only to stdout")
   # parser.add_argument('--sum', dest='accumulate', action='store_const',
   #                     const=sum, default=max,
   #                     help='sum the integers (default: find the max)')
@@ -23,13 +24,18 @@ def remove_qcow():
 
   if args.__contains__('pool'):
     settings.POOL = args.pool
+  if args.__contains__('noop'):
+    settings.NOOP = args.noop
 
   logger = logs.setup_logging()
 
   logger.info("deleting temp qcow for image %s from pool %s " % (image, settings.POOL))
 
   try:
-    os.remove("%s/%s/%s.qcow2" % (settings.QCOW_TEMP_PATH, settings.POOL, image))
+    if settings.NOOP:
+      logger.info('NOOP: would have removed temp qcow %s/%s/%s.qcow2' % (settings.QCOW_TEMP_PATH, settings.POOL, image))
+    else:
+      os.remove("%s/%s/%s.qcow2" % (settings.QCOW_TEMP_PATH, settings.POOL, image))
   except Exception as e:
     logger.error(e.stdout)
     logger.error(e.stderr)
